requirements.txt
Django
hubspot-api-client
djangorestframework
requests